feat: add order for following task forces
This commit modifies the code in the `taskForces.ts` file to add support for a new type of task force order: "follow". The code changes include adding a new type to the `TaskForceOrderMapper` and `TaskForceOrderInput` interfaces, as well as implementing the `TaskForceFollowOrder` resolver and updating the `tickTaskForceMovements` function to handle the new order type.

Signed-off-by: Jonathan Burke <jonathan.burke.1311@googlemail.com>